+++ src/vroom_int.h
@@ -30,7 +30,7 @@ Rcpp::IntegerVector read_int(vroom_vec_info* info) {
 
   parallel_for(
       n,
-      [&](int start, int end, int id) {
+      [&](size_t start, size_t end, size_t id) {
         size_t i = start;
         for (const auto& str :
              info->idx->get_column(info->column, start, end)) {
